moved distance function to diff file for testing
